requirements.txt
altgraph==0.17.4
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.3.2
cryptography==42.0.5
lxml==5.1.0
packaging==23.2
pdfminer.six==20221105
pdfplumber==0.10.4
pefile==2023.2.7
pillow==10.2.0
pycparser==2.21
pyinstaller==6.4.0
pyinstaller-hooks-contrib==2024.1
pypdfium2==4.27.0
python-docx==1.1.0
pywin32-ctypes==0.2.2
setuptools==69.1.1
typing_extensions==4.10.0
